Implement code review feedback
