+++ kie-dmn/kie-dmn-core/src/test/java/org/kie/dmn/core/compiler/profiles/ExtendedDMNProfileTest.java
@@ -3,20 +3,20 @@ package org.kie.dmn.core.compiler.profiles;
 import org.hamcrest.Matchers;
 import org.junit.Assert;
 import org.junit.Test;
+import org.kie.dmn.feel.runtime.functions.AbsFunction;
+import org.kie.dmn.feel.runtime.functions.EvenFunction;
+import org.kie.dmn.feel.runtime.functions.ExpFunction;
 import org.kie.dmn.feel.runtime.functions.FEELFnResult;
-import org.kie.dmn.feel.runtime.functions.extended.AbsFunction;
+import org.kie.dmn.feel.runtime.functions.LogFunction;
+import org.kie.dmn.feel.runtime.functions.MedianFunction;
+import org.kie.dmn.feel.runtime.functions.ModeFunction;
+import org.kie.dmn.feel.runtime.functions.ModuloFunction;
+import org.kie.dmn.feel.runtime.functions.OddFunction;
+import org.kie.dmn.feel.runtime.functions.ProductFunction;
+import org.kie.dmn.feel.runtime.functions.SplitFunction;
+import org.kie.dmn.feel.runtime.functions.SqrtFunction;
+import org.kie.dmn.feel.runtime.functions.StddevFunction;
 import org.kie.dmn.feel.runtime.functions.extended.DateFunction;
-import org.kie.dmn.feel.runtime.functions.extended.EvenFunction;
-import org.kie.dmn.feel.runtime.functions.extended.ExpFunction;
-import org.kie.dmn.feel.runtime.functions.extended.LogFunction;
-import org.kie.dmn.feel.runtime.functions.extended.MedianFunction;
-import org.kie.dmn.feel.runtime.functions.extended.ModeFunction;
-import org.kie.dmn.feel.runtime.functions.extended.ModuloFunction;
-import org.kie.dmn.feel.runtime.functions.extended.OddFunction;
-import org.kie.dmn.feel.runtime.functions.extended.ProductFunction;
-import org.kie.dmn.feel.runtime.functions.extended.SplitFunction;
-import org.kie.dmn.feel.runtime.functions.extended.SqrtFunction;
-import org.kie.dmn.feel.runtime.functions.extended.StddevFunction;
 import org.kie.dmn.feel.runtime.functions.extended.TimeFunction;
 
 import java.math.BigDecimal;
